8259845: Move placeholder implementation details to cpp file and add logging
