+++ src/hotspot/share/classfile/placeholders.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2003, 2020, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2003, 2021, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -25,10 +25,12 @@
 #ifndef SHARE_CLASSFILE_PLACEHOLDERS_HPP
 #define SHARE_CLASSFILE_PLACEHOLDERS_HPP
 
-#include "runtime/thread.hpp"
 #include "utilities/hashtable.hpp"
 
 class PlaceholderEntry;
+class Thread;
+class ClassLoaderData;
+class Symbol;
 
 // Placeholder objects. These represent classes currently
 // being loaded, as well as arrays of primitives.
@@ -95,44 +97,11 @@ public:
                        classloadAction action, Thread* thread);
 
   void print_on(outputStream* st) const;
+  void print() const;
   void verify();
 };
 
-// SeenThread objects represent list of threads that are
-// currently performing a load action on a class.
-// For class circularity, set before loading a superclass.
-// For bootclasssearchpath, set before calling load_instance_class.
-// Defining must be single threaded on a class/classloader basis
-// For DEFINE_CLASS, the head of the queue owns the
-// define token and the rest of the threads wait to return the
-// result the first thread gets.
-class SeenThread: public CHeapObj<mtInternal> {
-private:
-   Thread *_thread;
-   SeenThread* _stnext;
-   SeenThread* _stprev;
-public:
-   SeenThread(Thread *thread) {
-       _thread = thread;
-       _stnext = NULL;
-       _stprev = NULL;
-   }
-   Thread* thread()                const { return _thread;}
-   void set_thread(Thread *thread) { _thread = thread; }
-
-   SeenThread* next()              const { return _stnext;}
-   void set_next(SeenThread *seen) { _stnext = seen; }
-   void set_prev(SeenThread *seen) { _stprev = seen; }
-
-  void print_action_queue(outputStream* st) {
-    SeenThread* seen = this;
-    while (seen != NULL) {
-      seen->thread()->print_value_on(st);
-      st->print(", ");
-      seen = seen->next();
-    }
-  }
-};
+class SeenThread;
 
 // Placeholder objects represent classes currently being loaded.
 // All threads examining the placeholder table must hold the
@@ -141,6 +110,8 @@ public:
 // The system dictionary is the only user of this class.
 
 class PlaceholderEntry : public HashtableEntry<Symbol*, mtClass> {
+
+  friend class PlaceholderTable;
 
  private:
   ClassLoaderData*  _loader_data;   // initiating loader
@@ -156,6 +127,11 @@ class PlaceholderEntry : public HashtableEntry<Symbol*, mtClass> {
                                     // including _definer
                                     // _definer owns token
                                     // queue waits for and returns results from _definer
+
+  SeenThread* actionToQueue(PlaceholderTable::classloadAction action);
+  void set_threadQ(SeenThread* seenthread, PlaceholderTable::classloadAction action);
+  void add_seen_thread(Thread* thread, PlaceholderTable::classloadAction action);
+  bool remove_seen_thread(Thread* thread, PlaceholderTable::classloadAction action);
 
  public:
   // Simple accessors, used only by SystemDictionary
@@ -199,39 +175,6 @@ class PlaceholderEntry : public HashtableEntry<Symbol*, mtClass> {
     return (klassname() == class_name && loader_data() == loader);
   }
 
-  SeenThread* actionToQueue(PlaceholderTable::classloadAction action) {
-    SeenThread* queuehead = NULL;
-    switch (action) {
-      case PlaceholderTable::LOAD_INSTANCE:
-         queuehead = _loadInstanceThreadQ;
-         break;
-      case PlaceholderTable::LOAD_SUPER:
-         queuehead = _superThreadQ;
-         break;
-      case PlaceholderTable::DEFINE_CLASS:
-         queuehead = _defineThreadQ;
-         break;
-      default: Unimplemented();
-    }
-    return queuehead;
-  }
-
-  void set_threadQ(SeenThread* seenthread, PlaceholderTable::classloadAction action) {
-    switch (action) {
-      case PlaceholderTable::LOAD_INSTANCE:
-         _loadInstanceThreadQ = seenthread;
-         break;
-      case PlaceholderTable::LOAD_SUPER:
-         _superThreadQ = seenthread;
-         break;
-      case PlaceholderTable::DEFINE_CLASS:
-         _defineThreadQ = seenthread;
-         break;
-      default: Unimplemented();
-    }
-    return;
-  }
-
   bool super_load_in_progress() {
      return (_superThreadQ != NULL);
   }
@@ -244,70 +187,8 @@ class PlaceholderEntry : public HashtableEntry<Symbol*, mtClass> {
     return (_defineThreadQ != NULL);
   }
 
-// Doubly-linked list of Threads per action for class/classloader pair
-// Class circularity support: links in thread before loading superclass
-// bootstrapsearchpath support: links in a thread before load_instance_class
-// definers: use as queue of define requestors, including owner of
-// define token. Appends for debugging of requestor order
-  void add_seen_thread(Thread* thread, PlaceholderTable::classloadAction action) {
-    assert_lock_strong(SystemDictionary_lock);
-    SeenThread* threadEntry = new SeenThread(thread);
-    SeenThread* seen = actionToQueue(action);
-
-    if (seen == NULL) {
-      set_threadQ(threadEntry, action);
-      return;
-    }
-    SeenThread* next;
-    while ((next = seen->next()) != NULL) {
-      seen = next;
-    }
-    seen->set_next(threadEntry);
-    threadEntry->set_prev(seen);
-    return;
-  }
-
-  bool check_seen_thread(Thread* thread, PlaceholderTable::classloadAction action) {
-    assert_lock_strong(SystemDictionary_lock);
-    SeenThread* threadQ = actionToQueue(action);
-    SeenThread* seen = threadQ;
-    while (seen) {
-      if (thread == seen->thread()) {
-        return true;
-      }
-      seen = seen->next();
-    }
-    return false;
-  }
-
-  // returns true if seenthreadQ is now empty
-  // Note, caller must ensure probe still exists while holding
-  // SystemDictionary_lock
-  // ignores if cleanup has already been done
-  // if found, deletes SeenThread
-  bool remove_seen_thread(Thread* thread, PlaceholderTable::classloadAction action) {
-    assert_lock_strong(SystemDictionary_lock);
-    SeenThread* threadQ = actionToQueue(action);
-    SeenThread* seen = threadQ;
-    SeenThread* prev = NULL;
-    while (seen) {
-      if (thread == seen->thread()) {
-        if (prev) {
-          prev->set_next(seen->next());
-        } else {
-          set_threadQ(seen->next(), action);
-        }
-        if (seen->next()) {
-          seen->next()->set_prev(prev);
-        }
-        delete seen;
-        break;
-      }
-      prev = seen;
-      seen = seen->next();
-    }
-    return (actionToQueue(action) == NULL);
-  }
+  // Used for ClassCircularityError checking
+  bool check_seen_thread(Thread* thread, PlaceholderTable::classloadAction action);
 
   // Print method doesn't append a cr
   void print_entry(outputStream* st) const;
